Refactor user repository structure and add transactional unit of work
